requirements.txt
docker
pytenable